Extract ServicePrincipalCredentialCache
